Add EFI stub for arm64
This patch adds the EFI stub for arm64.  A PE/COFF header is added in head.S.
PE/COFF linker support is not available for arm64, so a native build is not
possible.  Also, this allows the binary to be both a PE/COFF EFI application,
and a normal Image file bootable like a Linux kernel.  The arm and arm64 Linux
kernels use the same methodology to create a single image that is both an EFI
application and a zImage/Image file.

The EFI stub processes the XEN EFI configuration file to load the dom0 kernel,
ramdisk, etc and constructs a device tree for XEN to use, then transfers
control to the normal XEN image entry point.  Device tree description of memory
is no longer used, as the stub uses the EFI memory map to populate bootinfo
memory banks.

Signed-off-by: Roy Franz <roy.franz@linaro.org>

Conflicts:

	xen/arch/arm/arm64/head.S